+++ src/utils/advancedSearchEnum.js
@@ -1,0 +1,6 @@
+export const AdvancedSearchEnum = {
+    SearchByName: 'SearchByName',
+    FindByMail: 'FindByMail',
+    FindByT: 'FindByT',
+}
+Object.freeze(AdvancedSearchEnum);