+++ CHANGELOG.md
@@ -5,11 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [unreleased]
+- minor: ([66](https://github.com/meateam/api-gateway/pull/66)): add advance search by mail/T.
+
 ## [v3.3.0] - 2021-03-14
 
 ### Added
 - minor([67](https://github.com/meateam/drive-ui/pull/67)): added cts 
 - minor([68](https://github.com/meateam/drive-ui/pull/68)): added button for local office.
+
 ## [v3.2.0] - 2021-02-28
 
 ### Fixed